Removed dependency to global $rdf
When including rdflib.js into a web application using webpack the
resulting application doesn't run. The reason for this is that there are
references to the global $rdf object in modules which are initialized
before the global $rdf object has been properly initialized.

Removed all references to the global $rdf object and
replaced them by direct references to the underlying functionality.